[RF] Remove dead code from HistFactory
* Remove the completely ununsed `HistFactorySimultaneous` (the regular
  RooSimultaneous is also used in HistFactory, and the
  `HistFactorySimultaneous` was just deriving from the RooSimultaneous
  without adding any new members.

* Remove the `src/Helper{.h,.cxx}` files, and move the functions from it
  that were actually used to the sources where they are used.

* Remove the `src/MakeModelAndMeasurements.cxx` source. The code was not
  used in any other sources, and the functions were unreachable because
  there was no corresponding header file.
