+++ roofit/histfactory/inc/LinkDef.h
@@ -24,7 +24,6 @@
 #pragma link C++ class RooStats::HistFactory::HistoToWorkspaceFactory+ ;
 #pragma link C++ class RooStats::HistFactory::HistoToWorkspaceFactoryFast+ ;
 #pragma link C++ class RooStats::HistFactory::RooBarlowBeestonLL+ ;
-#pragma link C++ class RooStats::HistFactory::HistFactorySimultaneous+ ;
 #pragma link C++ class RooStats::HistFactory::HistFactoryNavigation+ ;
 
 #pragma link C++ class RooStats::HistFactory::ConfigParser+ ;
