+++ roofit/histfactory/src/MakeModelAndMeasurementsFast.cxx
@@ -34,7 +34,6 @@
 #include "RooStats/ModelConfig.h"
 
 // from this package
-#include "Helper.h"
 #include "RooStats/HistFactory/EstimateSummary.h"
 #include "RooStats/HistFactory/Measurement.h"
 #include "RooStats/HistFactory/HistoToWorkspaceFactoryFast.h"
@@ -526,3 +525,174 @@ void RooStats::HistFactory::FormatFrameForLikelihood(RooPlot* frame, std::string
 }
 
 
+// Looking to deprecate this function and convert entirely to Measurements
+std::vector<RooStats::HistFactory::EstimateSummary> RooStats::HistFactory::GetChannelEstimateSummaries(
+        RooStats::HistFactory::Measurement& measurement,
+        RooStats::HistFactory::Channel& channel) {
+
+      // Convert a "Channel" into a list of "Estimate Summaries"
+      // This should only be a temporary function, as the
+      // EstimateSummary class should be deprecated
+
+      std::vector<EstimateSummary> channel_estimateSummary;
+
+      std::cout << "Processing data: " << std::endl;
+
+      // Add the data
+      EstimateSummary data_es;
+      data_es.name = "Data";
+      data_es.channel = channel.GetName();
+      TH1* data_hist = (TH1*) channel.GetData().GetHisto();
+      if( data_hist != NULL ) {
+   //data_es.nominal = (TH1*) channel.GetData().GetHisto()->Clone();
+   data_es.nominal = data_hist;
+   channel_estimateSummary.push_back( data_es );
+      }
+
+      // Add the samples
+      for( unsigned int sampleItr = 0; sampleItr < channel.GetSamples().size(); ++sampleItr ) {
+
+   EstimateSummary sample_es;
+   RooStats::HistFactory::Sample& sample = channel.GetSamples().at( sampleItr );
+
+   std::cout << "Processing sample: " << sample.GetName() << std::endl;
+
+   // Define the mapping
+   sample_es.name = sample.GetName();
+   sample_es.channel = sample.GetChannelName();
+   sample_es.nominal = (TH1*) sample.GetHisto()->Clone();
+
+   std::cout << "Checking NormalizeByTheory" << std::endl;
+
+   if( sample.GetNormalizeByTheory() ) {
+     sample_es.normName = "" ; // Really bad, confusion convention
+   }
+   else {
+     TString lumiStr;
+     lumiStr += measurement.GetLumi();
+     lumiStr.ReplaceAll(' ', TString());
+     sample_es.normName = lumiStr ;
+   }
+
+   std::cout << "Setting the Histo Systs" << std::endl;
+
+   // Set the Histo Systs:
+   for( unsigned int histoItr = 0; histoItr < sample.GetHistoSysList().size(); ++histoItr ) {
+
+     RooStats::HistFactory::HistoSys& histoSys = sample.GetHistoSysList().at( histoItr );
+
+     sample_es.systSourceForHist.push_back( histoSys.GetName() );
+     sample_es.lowHists.push_back( (TH1*) histoSys.GetHistoLow()->Clone()  );
+     sample_es.highHists.push_back( (TH1*) histoSys.GetHistoHigh()->Clone() );
+
+   }
+
+   std::cout << "Setting the NormFactors" << std::endl;
+
+   for( unsigned int normItr = 0; normItr < sample.GetNormFactorList().size(); ++normItr ) {
+
+     RooStats::HistFactory::NormFactor& normFactor = sample.GetNormFactorList().at( normItr );
+
+     EstimateSummary::NormFactor normFactor_es;
+     normFactor_es.name = normFactor.GetName();
+     normFactor_es.val  = normFactor.GetVal();
+     normFactor_es.high = normFactor.GetHigh();
+     normFactor_es.low  = normFactor.GetLow();
+     normFactor_es.constant = normFactor.GetConst();
+
+     sample_es.normFactor.push_back( normFactor_es );
+
+   }
+
+   std::cout << "Setting the OverallSysList" << std::endl;
+
+   for( unsigned int sysItr = 0; sysItr < sample.GetOverallSysList().size(); ++sysItr ) {
+
+     RooStats::HistFactory::OverallSys& overallSys = sample.GetOverallSysList().at( sysItr );
+
+     std::pair<double, double> DownUpPair( overallSys.GetLow(), overallSys.GetHigh() );
+     sample_es.overallSyst[ overallSys.GetName() ]  = DownUpPair; //
+
+   }
+
+   std::cout << "Checking Stat Errors" << std::endl;
+
+   // Do Stat Error
+   sample_es.IncludeStatError  = sample.GetStatError().GetActivate();
+
+   // Set the error and error threshold
+   sample_es.RelErrorThreshold = channel.GetStatErrorConfig().GetRelErrorThreshold();
+   if( sample.GetStatError().GetErrorHist() ) {
+     sample_es.relStatError      = (TH1*) sample.GetStatError().GetErrorHist()->Clone();
+   }
+   else {
+     sample_es.relStatError    = NULL;
+   }
+
+
+   // Set the constraint type;
+   Constraint::Type type = channel.GetStatErrorConfig().GetConstraintType();
+
+   // Set the default
+   sample_es.StatConstraintType = EstimateSummary::Gaussian;
+
+   if( type == Constraint::Gaussian) {
+     std::cout << "Using Gaussian StatErrors" << std::endl;
+     sample_es.StatConstraintType = EstimateSummary::Gaussian;
+   }
+   if( type == Constraint::Poisson ) {
+     std::cout << "Using Poisson StatErrors" << std::endl;
+     sample_es.StatConstraintType = EstimateSummary::Poisson;
+   }
+
+
+   std::cout << "Getting the shape Factor" << std::endl;
+
+   // Get the shape factor
+   if( sample.GetShapeFactorList().size() > 0 ) {
+     sample_es.shapeFactorName = sample.GetShapeFactorList().at(0).GetName();
+   }
+   if( sample.GetShapeFactorList().size() > 1 ) {
+     std::cout << "Error: Only One Shape Factor currently supported" << std::endl;
+     throw hf_exc();
+   }
+
+
+   std::cout << "Setting the ShapeSysts" << std::endl;
+
+   // Get the shape systs:
+   for( unsigned int shapeItr=0; shapeItr < sample.GetShapeSysList().size(); ++shapeItr ) {
+
+     RooStats::HistFactory::ShapeSys& shapeSys = sample.GetShapeSysList().at( shapeItr );
+
+     EstimateSummary::ShapeSys shapeSys_es;
+     shapeSys_es.name = shapeSys.GetName();
+     shapeSys_es.hist = shapeSys.GetErrorHist();
+
+     // Set the constraint type;
+     Constraint::Type systype = shapeSys.GetConstraintType();
+
+     // Set the default
+     shapeSys_es.constraint = EstimateSummary::Gaussian;
+
+     if( systype == Constraint::Gaussian) {
+       shapeSys_es.constraint = EstimateSummary::Gaussian;
+     }
+     if( systype == Constraint::Poisson ) {
+       shapeSys_es.constraint = EstimateSummary::Poisson;
+     }
+
+     sample_es.shapeSysts.push_back( shapeSys_es );
+
+   }
+
+   std::cout << "Adding this sample" << std::endl;
+
+   // Push back
+   channel_estimateSummary.push_back( sample_es );
+
+      }
+
+      return channel_estimateSummary;
+
+}
